Polish "Preserve nested Flyway migration paths in native images"
See gh-50433
